refactor ShopInventory name to ShopWindow
